Move IEvent to LogEvent conversion logics into extension method
